\name{print_lpa_summary}
\alias{print_lpa_summary}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Display information about the model.}
\description{Display information about the model.}
\usage{
print_lpa_summary(model, start_profile_labels_from1 = TRUE)
}
\arguments{
  \item{model}{A DPM-LPA model object.}
  \item{start_profile_labels_from1}{Whether to label latent profiles starting from 0 (Python style) or from 1 (R/Matlab/ordinary counting style). This only affects output labeling, not any calculations.}
}
\author{Samuel Paskewitz}
