\name{update_python_packages}
\alias{update_python_packages}
\title{
Update relevant Python packages.
}
\description{
Update relevant Python packages.
}
\usage{
update_python_packages()
}
%- maybe also 'usage' for other objects documented here.
\details{
Updates the Python packages 'numpy', 'scipy', 'pandas', 'vbayesfa', and
'plotnine' to the latest versions. 'vbayesfa' is the Python package that
contains the DPM-LPA code and the others are its dependencies.
}
\author{Sam Paskewitz}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
